This is the original Eagle Space Flight team communication system designed by Thomas Montano in 2018.
This will be considered the original ALPHA 1 release.

In the  documentation folder you will find a guide on how to use GNURADIO, the GRC files, my original
c++ code, and a writeup for how it should all operate.

Please note that this system was designed to operate using a BladeRF on the ground, and an ADALM Pluto
on the rocket.

-Thomas Montano 9/23/2018